requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-community
chromadb
sentence-transformers
langchain_huggingface
langchain-core
langchain_experimental
pydantic
python-multipart
typing-extensions
numpy
python-dotenv
PyPDF2
pandas
argon2-cffi
python-jose[cryptography]
aiofiles
uuid
typing
datetime
sqlalchemy
pydantic[email]
pathlib
hashlib
sqlite3
streamlit
requests